Completed Version 0 goals
By this commit, The integration is completed for two way communications
and basic functionality where from the qaul you can send Hii into the
matrix room and from matrix room you can send Hii into the qaul. To add
more, We have two commands running on matrix-side which are !qaul for
acknowledgement of connection and !users-list for getting list of users
in qaul on matrix.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>